requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.3.0
google-cloud-dialogflow==2.28.0
httpx==0.25.0
redis==5.0.1
python-multipart==0.0.6
tenacity==8.2.3
numpy==1.26.4
pandas==2.1.3
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-memcached==1.59
aiomysql==0.2.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
aiohttp=3.12.15